start: fix result check when using multiple domains
When using multiple domains (e.g. as a comma-separated list), certbot
will store the certificates in a folder named after the first domain.

Citing the docs:

"All domains will be included as Subject Alternative Names on the
certificate. The first domain will be used as the certificate name,
unless otherwise specified or if you already have a certificate with
the same name. In the case of a name conflict, a number like -0001
will be appended to the certificate name"